Refactor and reorganize classes
