+++ Assets/UGF.Testing.Runtime.Tests.Tests/Scopes/TestTestLogScope.cs
@@ -1,7 +1,8 @@
 using NUnit.Framework;
+using UGF.Testing.Runtime.Tests.Scopes;
 using UnityEngine;
 
-namespace UGF.Testing.Runtime.Tests.Tests
+namespace UGF.Testing.Runtime.Tests.Tests.Scopes
 {
     public class TestTestLogScope
     {
@@ -41,7 +42,7 @@ namespace UGF.Testing.Runtime.Tests.Tests
         [Test]
         public void Enabled()
         {
-            using (new TestLogScope(true))
+            using (new TestLogEnableScope(true))
             {
                 Debug.Log("Enabled()");
             }
@@ -52,7 +53,7 @@ namespace UGF.Testing.Runtime.Tests.Tests
         [Test]
         public void Disabled()
         {
-            using (new TestLogScope(false))
+            using (new TestLogEnableScope(false))
             {
                 Debug.Log("Disabled()");
             }
