+++ Assets/UGF.Testing.Runtime.Tests.Tests/Scopes/TestTestObjectScope.cs
@@ -1,8 +1,9 @@
 using NUnit.Framework;
+using UGF.Testing.Runtime.Tests.Scopes;
 using UnityEngine;
 using Object = UnityEngine.Object;
 
-namespace UGF.Testing.Runtime.Tests.Tests
+namespace UGF.Testing.Runtime.Tests.Tests.Scopes
 {
     public class TestTestObjectScope
     {
