+++ Assets/UGF.Testing.Runtime.Tests.Tests/Scopes/TestTestScriptableObjectScope.cs
@@ -1,7 +1,8 @@
 using NUnit.Framework;
+using UGF.Testing.Runtime.Tests.Scopes;
 using UnityEngine;
 
-namespace UGF.Testing.Runtime.Tests.Tests
+namespace UGF.Testing.Runtime.Tests.Tests.Scopes
 {
     public class TestTestScriptableObjectScope
     {
