+++ Packages/UGF.Testing/Runtime.Tests/Scopes/TestScriptableObjectScope.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace UGF.Testing.Runtime.Tests
+namespace UGF.Testing.Runtime.Tests.Scopes
 {
     /// <summary>
     /// Represents a scriptableObject scope.
