+++ Packages/UGF.Testing/Runtime.Tests/TestNoLogs.cs
@@ -7,12 +7,12 @@ namespace UGF.Testing.Runtime.Tests
     /// </summary>
     public abstract class TestNoLogs : TestBase
     {
-        public override void SetupAll()
+        public override void OnSetupAll()
         {
             Debug.unityLogger.logEnabled = false;
         }
 
-        public override void TeardownAll()
+        public override void OnTeardownAll()
         {
             Debug.unityLogger.logEnabled = true;
         }
